Add kielread EMU corpus data, which I use for training.
Following https://sourceforge.net/projects/emu/ is the EMU software published
under a BSD license, so I should be allowed to share these files which are part
of one of the examples included in that project, the 'kielread' corpus of spoken
read sentences in northern German. I shall try to find the appropriate
citations, attribution etc. for this data if I keep using it.